[Testing] Cleaned Up Test Scripts UI
The script to run the full AP flow did not provide many messages and did
not handle errors correctly, which made it challenging to debug. Fixed
these issues.

The AP run files were being saved into the intermediate files directory,
which made them hard to access and made it challenging to clear the
intermediate files. Moved them to their own directory.

Renamed the get_intermediate_file script to be plural.